LayoutTests/imported/w3c/web-platform-tests/credential-management/federatedcredential-framed-get.sub.https-expected.txt

Blocked access to external URL https://xn--lve-6lad.localhost:9443/credential-management/support/federatedcredential-get.html
Blocked access to external URL https://xn--lve-6lad.localhost:9443/credential-management/support/echoing-nester.html?origin=https://localhost:9443&file=federatedcredential-get.html
Blocked access to external URL https://xn--lve-6lad.localhost:9443/credential-management/support/echoing-nester.html?origin=https://xn--lve-6lad.localhost:9443&file=federatedcredential-get.html
Blocked access to external URL https://xn--lve-6lad.localhost:9443/credential-management/support/federatedcredential-get.html


Harness Error (TIMEOUT), message = null

FAIL Same-origin IFrame does not throw. assert_equals: expected "resolved" but got "rejected"
TIMEOUT Cross-origin IFrame throws 'NotAllowedError'. Test timed out
FAIL Same-origin IFrame in same-origin IFrame does not throw. assert_equals: expected "resolved" but got "rejected"
TIMEOUT Same-origin IFrame in same-origin IFrame throws 'NotAllowedError'. Test timed out
TIMEOUT Cross-origin IFrame in same-origin IFrame throws 'NotAllowedError'. Test timed out
TIMEOUT Cross-origin IFrame in same-cross-origin throws 'NotAllowedError'. Test timed out
